Move utils out of generators
